requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
array_record==0.8.1
asttokens==3.0.1
astunparse==1.6.3
attrs==25.3.0
av==15.1.0
beautifulsoup4==4.14.2
bitsandbytes==0.47.0
blessed==1.22.0
blinker==1.9.0
Brotli==1.1.0
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
bypy==1.8.9
cachetools==6.2.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1759648874697/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.3.0
cloudpickle==3.1.1
cmake==4.1.0
contourpy==1.3.3
crc32c==2.7.1
cycler==0.12.1
datasets==3.2.0
decorator==5.2.1
deepdiff==8.6.1
deepspeed==0.17.6
diffusers==0.34.0
dill==0.3.8
distro==1.9.0
dm-tree==0.1.9
docstring_parser==0.17.0
donfig==0.8.1.post1
dotenv==0.9.9
draccus==0.10.0
einops==0.8.1
et_xmlfile==2.0.0
etils==1.13.0
evaluate==0.4.6
evdev==1.9.2
executing==2.2.1
Farama-Notifications==0.0.4
fastapi==0.117.1
ffmpy==0.6.1
filelock @ file:///croot/filelock_1744281381737/work
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2024.9.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-auth==2.43.0
google-genai==1.51.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gpustat==1.1.1
gradio==5.47.1
gradio_client==1.13.2
groovy==0.1.2
grpcio==1.75.1
gymnasium==0.29.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.10
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
imageio-ffmpeg==0.6.0
immutabledict==4.2.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
inquirerpy==0.3.4
ipdb==0.13.13
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2 @ file:///croot/jinja2_1741710844255/work
jiter==0.11.1
joblib==1.5.2
json_repair==0.53.0
jsonlines==4.0.0
keras==3.11.3
kiwisolver==1.4.9
lazy_loader==0.4
-e git+ssh://git@github.com/paszea/lerobot.git@06442eff83ee78d00ea202283b98995b7cf29af7#egg=lerobot
libclang==18.1.1
lightning-utilities==0.15.2
llvmlite==0.45.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mergedeep==1.3.4
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml_dtypes==0.5.3
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.1
msgpack-numpy==0.4.8
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
namex==0.1.0
networkx @ file:///croot/networkx_1756709277056/work
ninja==1.13.0
nltk==3.9.2
numba==0.62.1
numcodecs==0.16.3
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==2.6.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
orderly-set==5.5.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pfzy==0.3.4
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1729065615156/work
piq==0.8.0
platformdirs==4.4.0
polars==1.35.1
polars-runtime-32==1.35.1
portalocker==3.2.0
promise==2.3
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser @ file:///home/task_175749606196006/conda-bld/pycparser_1757496076770/work
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pymunk==6.11.1
pynput==1.8.1
pyparsing==3.2.5
pyserial==3.5
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-xlib==0.33
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyyaml-include==1.4.1
pyzmq==27.1.0
qwen-vl-utils==0.0.14
regex==2025.9.18
requests @ file:///croot/requests_1756709366904/work
requests-toolbelt==1.0.0
rerun-sdk==0.22.1
rich==14.1.0
rsa==4.9.1
ruff==0.13.2
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.39.0
shellingham==1.5.4
shtab==1.7.2
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.48.0
sty==1.0.6
sympy @ file:///croot/sympy_1756713465822/work
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-datasets==4.9.9
tensorflow-metadata==1.17.2
termcolor==3.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
timeout-decorator==0.5.0
timm==1.0.20
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
torch-fidelity==0.3.0
torchaudio==2.3.0
torchcodec==0.5
torchmetrics==1.6.3
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.3.1
typeguard==4.4.4
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tyro==0.9.35
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1750775463400/work
uvicorn==0.37.0
validators==0.35.0
-e git+https://github.com/open-compass/VLMEvalKit.git@d4a3f652a3af51bdd8dedbd673455cab2612812d#egg=vlmeval
wandb==0.22.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xlsxwriter==3.2.9
xxhash==3.5.0
yarl==1.20.1
zarr==3.1.3
zipp==3.23.0
